requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.3
Django==5.2.7
django-ckeditor==6.7.3
django-jazzmin==3.0.1
django-js-asset==3.1.2
idna==3.10
pillow==11.3.0
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
